Unify spectroscopy input and connect test suites
